Refactor Time class and enhance multithreading support
- Changed static member variables in the Time class from double to float, updating GetDeltaTime method accordingly.
- Implemented multithreading in Application.cpp for improved performance, with proper locking mechanisms for thread safety.
- Updated Physics::Update to accept a reference to a vector, enhancing efficiency.
- Enabled V-Sync in Renderer by setting glfwSwapInterval to 1.

+++ Core/src/Time/Time.cpp
@@ -1,12 +1,12 @@
 #include "Time.h"
 
-double Time::mDeltaTime   = 0.0f;
-double Time::mCurrentTime = 0.0f;
-double Time::mLastSecond  = 0.0f;
-double Time::mLastTime    = 0.0f;
+float Time::mDeltaTime   = 0.0f;
+float Time::mCurrentTime = 0.0f;
+float Time::mLastSecond  = 0.0f;
+float Time::mLastTime    = 0.0f;
 int    Time::mFPS         = 0;
 
-double Time::GetDeltaTime()
+float Time::GetDeltaTime()
 {
     return mDeltaTime;
 }
