+++ Core/src/Time/Time.h
@@ -2,15 +2,15 @@
 #include "../ENGINE_API.h"
 class ENGINE_API Time
 {
-	static double mLastTime;
-	static double mCurrentTime;
-	static double mLastSecond;
-	static double mDeltaTime;
+	static float mLastTime;
+	static float mCurrentTime;
+	static float mLastSecond;
+	static float mDeltaTime;
 	static int mFPS;
 	friend class Application;
 	//static void SetDeltaTime(double time);
 public:
-	static double GetDeltaTime();
+	static float GetDeltaTime();
 	static int GetFPS();
 };
